+++ outputs.tf
@@ -513,6 +513,21 @@ output "vpc_endpoint_sqs_dns_entry" {
   value       = flatten(aws_vpc_endpoint.sqs.*.dns_entry)
 }
 
+output "vpc_endpoint_lambda_id" {
+  description = "The ID of VPC endpoint for Lambda"
+  value       = concat(aws_vpc_endpoint.lambda.*.id, [""])[0]
+}
+
+output "vpc_endpoint_lambda_network_interface_ids" {
+  description = "One or more network interfaces for the VPC Endpoint for Lambda."
+  value       = flatten(aws_vpc_endpoint.lambda.*.network_interface_ids)
+}
+
+output "vpc_endpoint_lambda_dns_entry" {
+  description = "The DNS entries for the VPC Endpoint for Lambda."
+  value       = flatten(aws_vpc_endpoint.lambda.*.dns_entry)
+}
+
 output "vpc_endpoint_codebuild_id" {
   description = "The ID of VPC endpoint for codebuild"
   value       = concat(aws_vpc_endpoint.codebuild.*.id, [""])[0]
@@ -1360,6 +1375,36 @@ output "vpc_endpoint_textract_network_interface_ids" {
 output "vpc_endpoint_textract_dns_entry" {
   description = "The DNS entries for the VPC Endpoint for Textract."
   value       = flatten(aws_vpc_endpoint.textract.*.dns_entry)
+}
+
+output "vpc_endpoint_codeartifact_api_id" {
+  description = "The ID of VPC endpoint for Codeartifact API"
+  value       = concat(aws_vpc_endpoint.codeartifact_api.*.id, [""])[0]
+}
+
+output "vpc_endpoint_codeartifact_api_network_interface_ids" {
+  description = "One or more network interfaces for the VPC Endpoint for Codeartifact API."
+  value       = flatten(aws_vpc_endpoint.codeartifact_api.*.network_interface_ids)
+}
+
+output "vpc_endpoint_codeartifact_api_dns_entry" {
+  description = "The DNS entries for the VPC Endpoint for Codeartifact API."
+  value       = flatten(aws_vpc_endpoint.codeartifact_api.*.dns_entry)
+}
+
+output "vpc_endpoint_codeartifact_repositories_id" {
+  description = "The ID of VPC endpoint for Codeartifact repositories"
+  value       = concat(aws_vpc_endpoint.codeartifact_repositories.*.id, [""])[0]
+}
+
+output "vpc_endpoint_codeartifact_repositories_network_interface_ids" {
+  description = "One or more network interfaces for the VPC Endpoint for Codeartifact repositories."
+  value       = flatten(aws_vpc_endpoint.codeartifact_repositories.*.network_interface_ids)
+}
+
+output "vpc_endpoint_codeartifact_repositories_dns_entry" {
+  description = "The DNS entries for the VPC Endpoint for Codeartifact repositories."
+  value       = flatten(aws_vpc_endpoint.codeartifact_repositories.*.dns_entry)
 }
 
 # VPC flow log